git content config moved to new file
